requirements.txt
diffusers==0.17.1
transformers==4.30.2
safetensors==0.3.1
ftfy==6.1.1
scipy==1.9.0
accelerate==0.18.0
bitsandbytes==0.37.2
numpy==1.24.1
opencv-python==4.7.0.68
tqdm==4.62.2
Pillow==9.4.0
timm==0.4.12
ipython==7.19.0
lingua-language-detector==1.3.2
huggingface-hub==0.15.1
boto3>=1.26.63,<2
boto3-type-annotations>=0.3.1,<1
responses==0.22.0
httpx==0.23.1
hypothesis==6.56.4
redis>=4,<5
python-dotenv
tabulate==0.9.0
pydantic==1.10.5
flask==2.2.3
waitress==2.1.2
nltk==3.8.1
pytorch-seed==0.2.0
pyloudnorm==0.1.1
git+https://github.com/suno-ai/bark.git
git+https://github.com/yekta/Kandinsky-2.git
git+https://github.com/openai/CLIP.git
git+https://github.com/yekta/denoiser.git
https://ba.stablecog.com/flash-attention/flash_attn-1.0.5-cp310-cp310-linux_x86_64.whl